Some quick n dirty code which generates pngs from some dicom files
Assumes a directory structure as described in the data/**/README.md
files. Running the dicom_utils script will generate a bunch of jpegs
corresponding to *some* of the dicom files it finds in the directory
tree.

It is still unable to load anything from the 1st dataset, I believe
because the image data is formatted differently. However, it does work
on most of the files in the second dataset.

The dicom files have a decent amount of other metadata in them as
well--I recommend loading up some of the files and looking through the
metadata fields.

A few definite TOODs from here:
* Come up with a file naming scheme/directory structure which will be
easy to consume with code and also easy to understand which images come
from the same patients/the same dates

* Fix the conversion failures

* Maybe make the dicom_utils script also output a csv of the non-image
metadata?